Move render function in controller
